Refactor to group functionality
