+++ app/input/types.py
@@ -1,0 +1,7 @@
+from enum import Enum, auto
+
+
+class InputType(Enum):
+    SCREEN = auto()
+    KEYBOARD = auto()
+    MOUSE = auto()
