+++ main.py
@@ -1,12 +1,8 @@
 import argparse
 from argparse import Namespace
 
-from app.input_listener import (
-    ScreenInputListener,
-    KeyboardInputListener,
-    MouseInputListener,
-)
-from app.storage_handler import FileSystemStorageHandler
+from app.input.core import ScreenInputListener, KeyboardInputListener, MouseInputListener
+from app.storage_handlers.core import FileSystemStorageHandler
 from app.record_keeper import RecordKeeper
